# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkillSync AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/skills:
    get:
      operationId: getUserSkills
      tags: [users]
      summary: Get user skills
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSkill"
    post:
      operationId: addUserSkill
      tags: [users]
      summary: Add a skill to user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSkillBody"
      responses:
        "201":
          description: Skill added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSkill"

  # JOBS
...
  /learning/progress:
    get:
      operationId: getLearningProgress
      tags: [learning]
      summary: Get learning progress and streaks
      responses:
        "200":
          description: Learning progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningProgress"
    post:
      operationId: updateLearningProgress
      tags: [learning]
      summary: Mark a learning item as complete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLearningProgressBody"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningProgress"

  # RESUME
  /resume:
    get:
      operationId: getResume
      tags: [resume]
      summary: Get current user resume info
      responses:
        "200":
          description: Resume info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    post:
      operationId: uploadResume
      tags: [resume]
      summary: Upload/create resume data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResumeBody"
      responses:
        "201":
          description: Resume created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resume/analyze:
    post:
      operationId: analyzeResume
      tags: [resume]
      summary: AI analysis of resume with ATS score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeResumeBody"
      responses:
        "200":
          description: Resume analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalysis"

  # CAREER
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, recruiter, faculty, freelancer_client, admin]
        avatar:
          type: string
        bio:
          type: string
        institution:
          type: string
        location:
          type: string
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        createdAt:
          type: string
      required: [id, name, email, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        institution:
          type: string
        location:
          type: string
        avatar:
          type: string
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [users, total, page, limit]
    UserSkill:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        skill:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        verified:
          type: boolean
      required: [id, userId, skill, level]
    AddSkillBody:
      type: object
      properties:
        skill:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
      required: [skill, level]

    # JOBS
...
    LearningProgress:
      type: object
      properties:
        streak:
          type: integer
        completedItems:
          type: integer
        totalItems:
          type: integer
        weeklyGoal:
          type: integer
        weeklyCompleted:
          type: integer
        completedIds:
          type: array
          items:
            type: integer
      required: [streak, completedItems, totalItems, weeklyGoal, weeklyCompleted, completedIds]
    UpdateLearningProgressBody:
      type: object
      properties:
        recommendationId:
          type: integer
      required: [recommendationId]

    # RESUME
    Resume:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        filename:
          type: string
        summary:
          type: string
        experience:
          type: array
          items:
            type: object
            properties:
              company:
                type: string
              role:
                type: string
              duration:
                type: string
            required: [company, role, duration]
        education:
          type: array
          items:
            type: object
            properties:
              institution:
                type: string
              degree:
                type: string
              year:
                type: string
            required: [institution, degree, year]
        extractedSkills:
          type: array
          items:
            type: string
        atsScore:
          type: number
        updatedAt:
          type: string
      required: [id, userId]
    CreateResumeBody:
      type: object
      properties:
        filename:
          type: string
        summary:
          type: string
        experience:
          type: array
          items:
            type: object
            properties:
              company:
                type: string
              role:
                type: string
              duration:
                type: string
            required: [company, role, duration]
        education:
          type: array
          items:
            type: object
            properties:
              institution:
                type: string
              degree:
                type: string
              year:
                type: string
            required: [institution, degree, year]
        skills:
          type: array
          items:
            type: string
      required: [filename]
    AnalyzeResumeBody:
      type: object
      properties:
        text:
          type: string
        targetJobTitle:
          type: string
      required: [text]
    ResumeAnalysis:
      type: object
      properties:
        atsScore:
          type: number
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        missingKeywords:
          type: array
          items:
            type: string
        extractedSkills:
          type: array
          items:
            type: string
      required: [atsScore, strengths, improvements, missingKeywords, extractedSkills]

    # CAREER
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, type, title, message, read, createdAt]

    # DASHBOARD
...